change stars to float type

+++ feedback/models.py
@@ -7,8 +7,8 @@ from django.core.validators import MinValueValidator, MaxValueValidator
 
 class Feedback(models.Model):
     message = models.TextField()
-    stars = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(5)], default=1)
+    stars = models.FloatField(validators=[MinValueValidator(1.0), MaxValueValidator(5.0)], default=1.0)  
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-
+    
     def __str__(self):
         return self.user.name
